moved root command and v1 root command to commands package - came up with a new entrypoint for go mains
