+++ cmd/gitsync/main.go
@@ -1,10 +1,10 @@
 package main
 
 import (
-	"github.com/restechnica/gitsync-cli/pkg/cli"
+	"github.com/restechnica/gitsync-cli/pkg/cli/exec"
 )
 
 // main runs the CLI.
 func main() {
-	_ = cli.Run()
+	_ = exec.Run()
 }
