+++ pkg/cli/exec/entrypoint.go
@@ -1,0 +1,8 @@
+package exec
+
+import "github.com/restechnica/gitsync-cli/pkg/cli/commands"
+
+func Run() error {
+	var command = commands.NewRootCommand()
+	return command.Execute()
+}
